move DiffEntry into submodule
